+++ tutorials/example_cpp_python_embedding_amar/func_4d.h
@@ -1,6 +1,6 @@
 #ifndef FUNC_4D_H
 #define FUNC_4D_H
 
-double func_4d(int x[4]);
+float func_4d(float x[4]);
 
 #endif
